add setDefaultAccount、optimize walletModel
